Split Task into multiple files.
This helps for readability, and keepign every file concerned with a
single aspect of 'Task'.
